Extract DataDecl and RType from Types.Types
