+++ liquidhaskell-boot/liquidhaskell-boot.cabal
@@ -87,6 +87,7 @@ library
                       Language.Haskell.Liquid.Transforms.InlineAux
                       Language.Haskell.Liquid.Types
                       Language.Haskell.Liquid.Types.Bounds
+                      Language.Haskell.Liquid.Types.DataDecl
                       Language.Haskell.Liquid.Types.Dictionaries
                       Language.Haskell.Liquid.Types.Equality
                       Language.Haskell.Liquid.Types.Errors
@@ -98,6 +99,7 @@ library
                       Language.Haskell.Liquid.Types.PredType
                       Language.Haskell.Liquid.Types.PrettyPrint
                       Language.Haskell.Liquid.Types.RefType
+                      Language.Haskell.Liquid.Types.RType
                       Language.Haskell.Liquid.Types.Specs
                       Language.Haskell.Liquid.Types.Types
                       Language.Haskell.Liquid.Types.Variance
