+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Types/Types.hs
@@ -21,7 +21,6 @@ module Language.Haskell.Liquid.Types.Types (
     module Language.Haskell.Liquid.UX.Config
 
   -- * Ghc Information
-  , TargetVars   (..)
   , TyConMap     (..)
 
   -- * F.Located Things
@@ -36,76 +35,20 @@ module Language.Haskell.Liquid.Types.Types (
   , F.dummyName
   , F.isDummy
 
-  -- * Bare Type Constructors and Variables
-  , BTyCon(..)
-  , mkBTyCon
-  -- , mkClassBTyCon, mkPromotedBTyCon
-  , isClassBTyCon
-  , BTyVar(..)
-
-  -- * Refined Type Constructors
-  , RTyCon (RTyCon, rtc_tc, rtc_info)
-  , TyConInfo(..), defaultTyConInfo
-  , rTyConPVs
-  , rTyConPropVs
-  -- , isClassRTyCon
-  , isClassType, isEqType, isRVar, isBool, isEmbeddedClass
-
   -- * Refinement Types
-  , RType (..), Ref(..), RTProp, rPropP
-  , RTyVar (..)
   , RTAlias (..)
   , OkRT
   , lmapEAlias
-
-  -- * Worlds
-  , HSeg (..)
-  , World (..)
 
   -- * Classes describing operations on `RTypes`
   , TyConable (..)
   , SubsTy (..)
 
-  -- * Type Variables
-  , RTVar (..), RTVInfo (..)
-  , makeRTVar, mapTyVarValue
-  , dropTyVarInfo, rTVarToBind
-  , setRtvPol
-
-  -- * Predicate Variables
-  , PVar (PV, pname, parg, ptype, pargs), isPropPV, pvType
-  , PVKind (..)
-  , Predicate (..)
-
-  -- * Refinements
-  , UReft(..)
-
   -- * Relational predicates
   , RelExpr (..)
 
-  -- * Parse-time entities describing refined data types
-  , SizeFun  (..), szFun
-  , DataDecl (..)
-  , DataName (..), dataNameSymbol
-  , DataCtor (..)
-  , DataConP (..)
-  , HasDataDecl (..), hasDecl
-  , DataDeclKind (..)
-  , TyConP   (..)
-
   -- * Pre-instantiated RType
-  , RRType, RRProp
-  , BRType, BRProp
-  , BSort, BPVar
-  , RTVU, PVU
-
-  -- * Instantiated RType
-  , BareType
-  , SpecType, SpecProp, SpecRTVar
   , SpecRep
-  , LocBareType, LocSpecType
-  , RSort
-  , UsedPVar, RPVar, RReft
   , REnv
   , AREnv (..)
 
@@ -116,7 +59,7 @@ module Language.Haskell.Liquid.Types.Types (
   , rFun, rFun', rCls, rRCls, rFunDebug
 
   -- * Manipulating `Predicates`
-  , pvars, pappSym, pApp
+  , pappSym, pApp
 
   -- * Some tests on RTypes
   , isBase
@@ -161,11 +104,6 @@ module Language.Haskell.Liquid.Types.Types (
   , F.PPrint (..)
   , F.pprint
   , F.showpp
-
-  -- * Printer Configuration
-  , PPEnv (..)
-  , ppEnv
-  , ppEnvShort
 
   -- * Modules and Imports
   , ModName (..), ModType (..)
@@ -234,13 +172,12 @@ module Language.Haskell.Liquid.Types.Types (
 
   , Axiom(..), HAxiom
 
-  -- , rtyVarUniqueSymbol, tyVarUniqueSymbol
-  , rtyVarType, tyVarVar
-
   -- * Refined Function Info
-  , RFInfo(..), defRFInfo, mkRFInfo, classRFInfo, classRFInfoType
+  , classRFInfoType
 
   , ordSrcSpan
+  , module Language.Haskell.Liquid.Types.DataDecl
+  , module Language.Haskell.Liquid.Types.RType
   )
   where
 
@@ -279,25 +216,22 @@ import           Data.Bifunctor
 import           Data.Typeable                          (Typeable)
 import           Data.Generics                          (Data)
 import qualified Data.Binary                            as B
-import qualified Data.Foldable                          as F
 import           Data.Hashable
 import qualified Data.HashMap.Strict                    as M
 import qualified Data.HashSet                           as S
-import qualified Data.List                              as L
-import           Data.Maybe                             (mapMaybe)
 import           Data.Function                          (on)
-import           Data.List                              as L (nub, null)
 import           Data.Text                              (Text)
 import           Text.PrettyPrint.HughesPJ              hiding (first, (<>))
-import           Text.Printf
 import           Language.Fixpoint.Misc
 
 import qualified Language.Fixpoint.Types as F
+import           Language.Fixpoint.Types (Expr, Symbol)
 
 import           Language.Haskell.Liquid.GHC.Misc
 import           Language.Haskell.Liquid.GHC.Logging as GHC
-import           Language.Haskell.Liquid.Types.Variance
+import           Language.Haskell.Liquid.Types.DataDecl
 import           Language.Haskell.Liquid.Types.Errors
+import           Language.Haskell.Liquid.Types.RType
 import           Language.Haskell.Liquid.Misc
 import           Language.Haskell.Liquid.UX.Config
 
@@ -320,62 +254,49 @@ data TyConMap = TyConMap
   , tcmFtcArity :: M.HashMap TyCon             Int     -- ^ Arity of each Family-Tycon
   }
 
-
-newtype RFInfo = RFInfo {permitTC :: Maybe Bool }
-  deriving (Generic, Data, Typeable, Show, Eq)
-
-defRFInfo :: RFInfo
-defRFInfo = RFInfo Nothing
-
-classRFInfo :: Bool -> RFInfo
-classRFInfo b = RFInfo $ Just b
-
 classRFInfoType :: Bool -> RType c tv r -> RType c tv r
 classRFInfoType b = fromRTypeRep .
                     (\trep@RTypeRep{..} -> trep{ty_info = map (\i -> i{permitTC = pure b}) ty_info}) .
                     toRTypeRep
 
-mkRFInfo :: Config  -> RFInfo
-mkRFInfo cfg = RFInfo $ Just (typeclass cfg)
-
-instance Hashable RFInfo
-instance NFData RFInfo
-instance B.Binary RFInfo
-
 -----------------------------------------------------------------------------
--- | Printer ----------------------------------------------------------------
+-- | Relational predicates --------------------------------------------------
 -----------------------------------------------------------------------------
 
-data PPEnv = PP
-  { ppPs    :: Bool -- ^ print abstract-predicates
-  , ppTyVar :: Bool -- ^ print the unique suffix for each tyvar
-  , ppShort :: Bool -- ^ print the tycons without qualification
-  , ppDebug :: Bool -- ^ gross with full info
-  }
-  deriving (Show)
+data RelExpr = ERBasic F.Expr | ERChecked Expr RelExpr | ERUnChecked Expr RelExpr
+  deriving (Eq, Show, Data, Generic)
 
-ppEnv :: PPEnv
-ppEnv = ppEnvDef
-          { ppPs    = True }
-          { ppDebug = True }   -- RJ: needed for resolution, because pp is used for serialization?
+instance B.Binary RelExpr
 
-{- | [NOTE:ppEnv] For some mysterious reason, `ppDebug` must equal `True`
-     or various tests fail e.g. tests/classes/pos/TypeEquality0{0,1}.hs
-     Yikes. Find out why!
- -}
+instance F.PPrint RelExpr where
+  pprintTidy k (ERBasic e)       = F.pprintTidy k e
+  pprintTidy k (ERChecked e r)   = F.pprintTidy k e <+> "!=>" <+> F.pprintTidy k r
+  pprintTidy k (ERUnChecked e r) = F.pprintTidy k e <+> ":=>" <+> F.pprintTidy k r
 
-ppEnvDef :: PPEnv
-ppEnvDef = PP False False False False
 
-ppEnvShort :: PPEnv -> PPEnv
-ppEnvShort pp = pp { ppShort = True }
+ignoreOblig :: RType t t1 t2 -> RType t t1 t2
+ignoreOblig (RRTy _ _ _ t) = t
+ignoreOblig t              = t
 
-------------------------------------------------------------------
--- Huh?
-------------------------------------------------------------------
-type Expr      = F.Expr
-type Symbol    = F.Symbol
 
+type SpecRTEnv   = RTEnv RTyVar SpecType
+type BareRTEnv   = RTEnv Symbol BareType
+type BareRTAlias = RTAlias Symbol BareType
+type SpecRTAlias = RTAlias RTyVar SpecType
+
+
+class SubsTy tv ty a where
+  subt :: (tv, ty) -> a -> a
+
+-- Should just make this a @Pretty@ instance but its too damn tedious
+-- to figure out all the constraints.
+
+type OkRT c tv r = ( TyConable c
+                   , F.PPrint tv, F.PPrint c, F.PPrint r
+                   , Reftable r, Reftable (RTProp c tv ()), Reftable (RTProp c tv r)
+                   , Eq c, Eq tv
+                   , Hashable tv
+                   )
 
 -- [NOTE:LIFTED-VAR-SYMBOLS]: Following NOTE:REFLECT-IMPORTS, by default
 -- each (lifted) `Var` is mapped to its `Symbol` via the `Symbolic Var`
@@ -433,665 +354,6 @@ isApp (F.EApp (F.EVar _) (F.EVar _)) = True
 isApp (F.EApp e (F.EVar _))          = isApp e
 isApp _                              = False
 
-data TyConP = TyConP
-  { tcpLoc          :: !F.SourcePos
-  , tcpCon          :: !TyCon
-  , tcpFreeTyVarsTy :: ![RTyVar]
-  , tcpFreePredTy   :: ![PVar RSort]
-  , tcpVarianceTs   :: !VarianceInfo
-  , tcpVariancePs   :: !VarianceInfo
-  , tcpSizeFun      :: !(Maybe SizeFun)
-  } deriving (Generic, Data, Typeable)
-
-instance F.Loc TyConP where
-  srcSpan tc = F.SS (tcpLoc tc) (tcpLoc tc)
-
-
--- TODO: just use Located instead of dc_loc, dc_locE
-data DataConP = DataConP
-  { dcpLoc        :: !F.SourcePos
-  , dcpCon        :: !DataCon                -- ^ Corresponding GHC DataCon
-  , dcpFreeTyVars :: ![RTyVar]               -- ^ Type parameters
-  , dcpFreePred   :: ![PVar RSort]           -- ^ Abstract Refinement parameters
-  , dcpTyConstrs  :: ![SpecType]             -- ^ ? Class constraints (via `dataConStupidTheta`)
-  , dcpTyArgs     :: ![(Symbol, SpecType)]   -- ^ Value parameters
-  , dcpTyRes      :: !SpecType               -- ^ Result type
-  , dcpIsGadt     :: !Bool                   -- ^ Was this specified in GADT style (if so, DONT use function names as fields)
-  , dcpModule     :: !F.Symbol               -- ^ Which module was this defined in
-  , dcpLocE       :: !F.SourcePos
-  } deriving (Generic, Data, Typeable)
-
--- | [NOTE:DataCon-Data] for each 'DataConP' we also
---   store the type of the constructed data. This is
---   *the same as* 'tyRes' for *vanilla* ADTs
---   (e.g. List, Maybe etc.) but may differ for GADTs.
---   For example,
---
---      data Thing a where
---        X  :: Thing Int
---        Y  :: Thing Bool
---
---   Here the 'DataConP' associated with 'X' (resp. 'Y')
---   has 'tyRes' corresponding to 'Thing Int' (resp. 'Thing Bool'),
---   but in both cases, the 'tyData' should be 'Thing a'.
---
-
-instance F.Loc DataConP where
-  srcSpan d = F.SS (dcpLoc d) (dcpLocE d)
-
--- | Which Top-Level Binders Should be Verified
-data TargetVars = AllVars | Only ![Var]
-
-
---------------------------------------------------------------------
--- | Abstract Predicate Variables ----------------------------------
---------------------------------------------------------------------
-
-data PVar t = PV
-  { pname :: !Symbol
-  , ptype :: !(PVKind t)
-  , parg  :: !Symbol
-  , pargs :: ![(t, Symbol, Expr)]
-  } deriving (Generic, Data, Typeable, Show, Functor)
-
-instance Eq (PVar t) where
-  pv == pv' = pname pv == pname pv' {- UNIFY: What about: && eqArgs pv pv' -}
-
-instance Ord (PVar t) where
-  compare (PV n _ _ _)  (PV n' _ _ _) = compare n n'
-
-instance B.Binary t => B.Binary (PVar t)
-instance NFData t   => NFData   (PVar t)
-
-instance Hashable (PVar a) where
-  hashWithSalt i (PV n _ _ _) = hashWithSalt i n
-
-pvType :: PVar t -> t
-pvType p = case ptype p of
-             PVProp t -> t
-             PVHProp  -> panic Nothing "pvType on HProp-PVar"
-
-data PVKind t
-  = PVProp t
-  | PVHProp
-  deriving (Generic, Data, Typeable, Functor, F.Foldable, Traversable, Show)
-
-instance B.Binary a => B.Binary (PVKind a)
-instance NFData a   => NFData   (PVKind a)
-
-
---------------------------------------------------------------------------------
--- | Predicates ----------------------------------------------------------------
---------------------------------------------------------------------------------
-
-type UsedPVar      = PVar ()
-
-newtype Predicate  = Pr [UsedPVar]
-  deriving (Generic, Data, Typeable)
-  deriving Hashable via Generically Predicate
-
-instance Eq Predicate where
-  (Pr vs) == (Pr ws)
-      = and $ (length vs' == length ws') : [v == w | (v, w) <- zip vs' ws']
-        where
-          vs' = L.sort vs
-          ws' = L.sort ws
-
-
-
-instance B.Binary Predicate
-
-instance NFData Predicate where
-  rnf _ = ()
-
-instance Monoid Predicate where
-  mempty  = pdTrue
-  mappend = (<>)
-
-instance Semigroup Predicate where
-  p <> p' = pdAnd [p, p']
-
-instance Semigroup a => Semigroup (UReft a) where
-  MkUReft x y <> MkUReft x' y' = MkUReft (x <> x') (y <> y')
-
-instance (Monoid a) => Monoid (UReft a) where
-  mempty  = MkUReft mempty mempty
-  mappend = (<>)
-
-
-pdTrue :: Predicate
-pdTrue         = Pr []
-
-pdAnd :: Foldable t => t Predicate -> Predicate
-pdAnd ps       = Pr (nub $ concatMap pvars ps)
-
-pvars :: Predicate -> [UsedPVar]
-pvars (Pr pvs) = pvs
-
-instance F.Subable UsedPVar where
-  syms pv         = [ y | (_, x, F.EVar y) <- pargs pv, x /= y ]
-  subst s pv      = pv { pargs = mapThd3 (F.subst s)  <$> pargs pv }
-  substf f pv     = pv { pargs = mapThd3 (F.substf f) <$> pargs pv }
-  substa f pv     = pv { pargs = mapThd3 (F.substa f) <$> pargs pv }
-
-
-instance F.Subable Predicate where
-  syms     (Pr pvs) = concatMap F.syms   pvs
-  subst  s (Pr pvs) = Pr (F.subst s  <$> pvs)
-  substf f (Pr pvs) = Pr (F.substf f <$> pvs)
-  substa f (Pr pvs) = Pr (F.substa f <$> pvs)
-
-instance NFData r => NFData (UReft r)
-
-data RelExpr = ERBasic F.Expr | ERChecked Expr RelExpr | ERUnChecked Expr RelExpr
-  deriving (Eq, Show, Data, Generic)
-
-instance B.Binary RelExpr
-
-instance F.PPrint RelExpr where
-  pprintTidy k (ERBasic e)       = F.pprintTidy k e
-  pprintTidy k (ERChecked e r)   = F.pprintTidy k e <+> "!=>" <+> F.pprintTidy k r
-  pprintTidy k (ERUnChecked e r) = F.pprintTidy k e <+> ":=>" <+> F.pprintTidy k r
-
-newtype BTyVar = BTV Symbol deriving (Show, Generic, Data, Typeable)
-
-newtype RTyVar = RTV TyVar deriving (Generic, Data, Typeable)
-
-instance Eq BTyVar where
-  (BTV x) == (BTV y) = x == y
-
-instance Ord BTyVar where
-  compare (BTV x) (BTV y) = compare x y
-
-instance IsString BTyVar where
-  fromString = BTV . fromString
-
-instance B.Binary BTyVar
-instance Hashable BTyVar
-instance NFData   BTyVar
-instance NFData   RTyVar
-
-instance F.Symbolic BTyVar where
-  symbol (BTV tv) = tv
-
-instance F.Symbolic RTyVar where
-  symbol (RTV tv) = F.symbol tv -- tyVarUniqueSymbol tv
-
--- instance F.Symbolic RTyVar where
-  -- symbol (RTV tv) = F.symbol . getName $ tv
--- rtyVarUniqueSymbol  :: RTyVar -> Symbol
--- rtyVarUniqueSymbol (RTV tv) = tyVarUniqueSymbol tv
--- tyVarUniqueSymbol :: TyVar -> Symbol
--- tyVarUniqueSymbol tv = F.symbol $ show (getName tv) ++ "_" ++ show (varUnique tv)
-
-data BTyCon = BTyCon
-  { btc_tc    :: !F.LocSymbol    -- ^ TyCon name with location information
-  , btc_class :: !Bool           -- ^ Is this a class type constructor?
-  , btc_prom  :: !Bool           -- ^ Is Promoted Data Con?
-  }
-  deriving (Generic, Data, Typeable)
-  deriving Hashable via Generically BTyCon
-
-instance B.Binary BTyCon
-
-data RTyCon = RTyCon
-  { rtc_tc    :: TyCon         -- ^ GHC Type Constructor
-  , rtc_pvars :: ![RPVar]      -- ^ Predicate Parameters
-  , rtc_info  :: !TyConInfo    -- ^ TyConInfo
-  }
-  deriving (Generic, Data, Typeable)
-
-instance F.Symbolic RTyCon where
-  symbol = F.symbol . rtc_tc
-
-instance F.Symbolic BTyCon where
-  symbol = F.val . btc_tc
-
-instance NFData BTyCon
-
-instance NFData RTyCon
-
-rtyVarType :: RTyVar -> Type
-rtyVarType (RTV v) = TyVarTy v
-
-tyVarVar :: RTVar RTyVar c -> Var
-tyVarVar (RTVar (RTV v) _) = v
-
-
-
-mkBTyCon :: F.LocSymbol -> BTyCon
-mkBTyCon x = BTyCon x False False
-
-
--- | Accessors for @RTyCon@
-
-isBool :: RType RTyCon t t1 -> Bool
-isBool (RApp RTyCon{rtc_tc = c} _ _ _) = c == boolTyCon
-isBool _                                 = False
-
-isRVar :: RType c tv r -> Bool
-isRVar (RVar _ _) = True
-isRVar _          = False
-
-isClassBTyCon :: BTyCon -> Bool
-isClassBTyCon = btc_class
-
--- isClassRTyCon :: RTyCon -> Bool
--- isClassRTyCon x = (isClassTyCon $ rtc_tc x) || (rtc_tc x == eqPrimTyCon)
-
-rTyConPVs :: RTyCon -> [RPVar]
-rTyConPVs     = rtc_pvars
-
-rTyConPropVs :: RTyCon -> [PVar RSort]
-rTyConPropVs  = filter isPropPV . rtc_pvars
-
-isPropPV :: PVar t -> Bool
-isPropPV      = isProp . ptype
-
-isEqType :: TyConable c => RType c t t1 -> Bool
-isEqType (RApp c _ _ _) = isEqual c
-isEqType _              = False
-
-
-isClassType :: TyConable c => RType c t t1 -> Bool
-isClassType (RApp c _ _ _) = isClass c
-isClassType _              = False
-
-isEmbeddedClass :: TyConable c => RType c t t1 -> Bool
-isEmbeddedClass (RApp c _ _ _) = isEmbeddedDict c
-isEmbeddedClass _              = False
-
--- rTyConPVHPs = filter isHPropPV . rtc_pvars
--- isHPropPV   = not . isPropPV
-
-isProp :: PVKind t -> Bool
-isProp (PVProp _) = True
-isProp _          = False
-
-
-defaultTyConInfo :: TyConInfo
-defaultTyConInfo = TyConInfo [] [] Nothing
-
-
------------------------------------------------------------------------
--- | Co- and Contra-variance for TyCon --------------------------------
------------------------------------------------------------------------
-
--- | Indexes start from 0 and type or predicate arguments can be both
---   covariant and contravaariant e.g., for the below Foo dataType
---
---     data Foo a b c d <p :: b -> Prop, q :: Int -> Prop, r :: a -> Prop>
---       = F (a<r> -> b<p>) | Q (c -> a) | G (Int<q> -> a<r>)
---
---  there will be:
---
---    varianceTyArgs     = [Bivariant , Covariant, Contravatiant, Invariant]
---    variancePsArgs     = [Covariant, Contravatiant, Bivariant]
---
-
-data TyConInfo = TyConInfo
-  { varianceTyArgs  :: !VarianceInfo      -- ^ variance info for type variables
-  , variancePsArgs  :: !VarianceInfo      -- ^ variance info for predicate variables
-  , sizeFunction    :: !(Maybe SizeFun)   -- ^ logical UNARY function that computes the size of the structure
-  } deriving (Generic, Data, Typeable)
-
-instance NFData TyConInfo
-
-instance Show TyConInfo where
-  show (TyConInfo x y _) = show x ++ "\n" ++ show y
-
---------------------------------------------------------------------------------
--- | Unified Representation of Refinement Types --------------------------------
---------------------------------------------------------------------------------
-
-type RTVU c tv = RTVar tv (RType c tv ())
-type PVU  c tv = PVar     (RType c tv ())
-
-instance Show tv => Show (RTVU c tv) where
-  show (RTVar t _) = show t
-
-data RType c tv r
-  = RVar {
-      rt_var    :: !tv
-    , rt_reft   :: !r
-    }
-
-  | RFun  {
-      rt_bind   :: !Symbol
-    , rt_rinfo  :: !RFInfo
-    , rt_in     :: !(RType c tv r)
-    , rt_out    :: !(RType c tv r)
-    , rt_reft   :: !r
-    }
-
-  | RAllT {
-      rt_tvbind :: !(RTVU c tv) -- RTVar tv (RType c tv ()))
-    , rt_ty     :: !(RType c tv r)
-    , rt_ref    :: !r
-    }
-
-  -- | "forall x y <z :: Nat, w :: Int> . TYPE"
-  --               ^^^^^^^^^^^^^^^^^^^ (rt_pvbind)
-  | RAllP {
-      rt_pvbind :: !(PVU c tv)
-    , rt_ty     :: !(RType c tv r)
-    }
-
-  -- | For example, in [a]<{\h -> v > h}>, we apply (via `RApp`)
-  --   * the `RProp`  denoted by `{\h -> v > h}` to
-  --   * the `RTyCon` denoted by `[]`.
-  | RApp  {
-      rt_tycon  :: !c
-    , rt_args   :: ![RType  c tv r]
-    , rt_pargs  :: ![RTProp c tv r]
-    , rt_reft   :: !r
-    }
-
-  | RAllE {
-      rt_bind   :: !Symbol
-    , rt_allarg :: !(RType c tv r)
-    , rt_ty     :: !(RType c tv r)
-    }
-
-  | REx {
-      rt_bind   :: !Symbol
-    , rt_exarg  :: !(RType c tv r)
-    , rt_ty     :: !(RType c tv r)
-    }
-
-  | RExprArg (F.Located Expr)                   -- ^ For expression arguments to type aliases
-                                                --   see tests/pos/vector2.hs
-  | RAppTy{
-      rt_arg   :: !(RType c tv r)
-    , rt_res   :: !(RType c tv r)
-    , rt_reft  :: !r
-    }
-
-  | RRTy  {
-      rt_env   :: ![(Symbol, RType c tv r)]
-    , rt_ref   :: !r
-    , rt_obl   :: !Oblig
-    , rt_ty    :: !(RType c tv r)
-    }
-
-  | RHole r -- ^ let LH match against the Haskell type and add k-vars, e.g. `x:_`
-            --   see tests/pos/Holes.hs
-  deriving (Eq, Generic, Data, Typeable, Functor)
-  deriving Hashable via Generically (RType c tv r)
-
-instance (B.Binary c, B.Binary tv, B.Binary r) => B.Binary (RType c tv r)
-instance (NFData c, NFData tv, NFData r)       => NFData (RType c tv r)
-
-ignoreOblig :: RType t t1 t2 -> RType t t1 t2
-ignoreOblig (RRTy _ _ _ t) = t
-ignoreOblig t              = t
-
-makeRTVar :: tv -> RTVar tv s
-makeRTVar a = RTVar a (RTVNoInfo True)
-
-instance (Eq tv) => Eq (RTVar tv s) where
-  t1 == t2 = ty_var_value t1 == ty_var_value t2
-
-data RTVar tv s = RTVar
-  { ty_var_value :: tv
-  , ty_var_info  :: RTVInfo s
-  } deriving (Generic, Data, Typeable)
-    deriving Hashable via Generically (RTVar tv s)
-
-mapTyVarValue :: (tv1 -> tv2) -> RTVar tv1 s -> RTVar tv2 s
-mapTyVarValue f v = v {ty_var_value = f $ ty_var_value v}
-
-dropTyVarInfo :: RTVar tv s1 -> RTVar tv s2
-dropTyVarInfo v = v{ty_var_info = RTVNoInfo True }
-
-data RTVInfo s
-  = RTVNoInfo { rtv_is_pol :: Bool }
-  | RTVInfo { rtv_name   :: Symbol
-            , rtv_kind   :: s
-            , rtv_is_val :: Bool
-            , rtv_is_pol :: Bool -- true iff the type variable gets instantiated with
-                                 -- any refinement (ie is polymorphic on refinements),
-                                 -- false iff instantiation is with true refinement
-            } deriving (Generic, Data, Typeable, Functor, Eq)
-              deriving Hashable via Generically (RTVInfo s)
-
-
-setRtvPol :: RTVar tv a -> Bool -> RTVar tv a
-setRtvPol (RTVar a i) b = RTVar a (i{rtv_is_pol = b})
-
-rTVarToBind :: RTVar RTyVar s  -> Maybe (Symbol, s)
-rTVarToBind = go . ty_var_info
-  where
-    go RTVInfo{..} | rtv_is_val = Just (rtv_name, rtv_kind)
-    go _                        = Nothing
-
-tyVarIsVal :: RTVar tv s -> Bool
-tyVarIsVal = rtvinfoIsVal . ty_var_info
-
-rtvinfoIsVal :: RTVInfo s -> Bool
-rtvinfoIsVal RTVNoInfo{} = False
-rtvinfoIsVal RTVInfo{..} = rtv_is_val
-
-instance (B.Binary tv, B.Binary s) => B.Binary (RTVar tv s)
-instance (NFData tv, NFData s)     => NFData   (RTVar tv s)
-instance (NFData s)                => NFData   (RTVInfo s)
-instance (B.Binary s)              => B.Binary (RTVInfo s)
-
--- | @Ref@ describes `Prop τ` and `HProp` arguments applied to type constructors.
---   For example, in [a]<{\h -> v > h}>, we apply (via `RApp`)
---   * the `RProp`  denoted by `{\h -> v > h}` to
---   * the `RTyCon` denoted by `[]`.
---   Thus, @Ref@ is used for abstract-predicate (arguments) that are associated
---   with _type constructors_ i.e. whose semantics are _dependent upon_ the data-type.
---   In contrast, the `Predicate` argument in `ur_pred` in the @UReft@ applies
---   directly to any type and has semantics _independent of_ the data-type.
-
-data Ref τ t = RProp
-  { rf_args :: [(Symbol, τ)] -- ^ arguments. e.g. @h@ in the above example
-  , rf_body :: t -- ^ Abstract refinement associated with `RTyCon`. e.g. @v > h@ in the above example
-  } deriving (Eq, Generic, Data, Typeable, Functor)
-    deriving Hashable via Generically (Ref τ t)
-
-instance (B.Binary τ, B.Binary t) => B.Binary (Ref τ t)
-instance (NFData τ,   NFData t)   => NFData   (Ref τ t)
-
-rPropP :: [(Symbol, τ)] -> r -> Ref τ (RType c tv r)
-rPropP τ r = RProp τ (RHole r)
-
--- | @RTProp@ is a convenient alias for @Ref@ that will save a bunch of typing.
---   In general, perhaps we need not expose @Ref@ directly at all.
-type RTProp c tv r = Ref (RType c tv ()) (RType c tv r)
-
-
--- | A @World@ is a Separation Logic predicate that is essentially a sequence of binders
---   that satisfies two invariants (TODO:LIQUID):
---   1. Each `hs_addr :: Symbol` appears at most once,
---   2. There is at most one `HVar` in a list.
-
-newtype World t = World [HSeg t]
-                deriving (Generic, Data, Typeable)
-
-data    HSeg  t = HBind {hs_addr :: !Symbol, hs_val :: t}
-                | HVar UsedPVar
-                deriving (Generic, Data, Typeable)
-
-data UReft r = MkUReft
-  { ur_reft   :: !r
-  , ur_pred   :: !Predicate
-  }
-  deriving (Eq, Generic, Data, Typeable, Functor, Foldable, Traversable)
-  deriving Hashable via Generically (UReft r)
-
-instance B.Binary r => B.Binary (UReft r)
-
-type BRType      = RType BTyCon BTyVar       -- ^ "Bare" parsed version
-type RRType      = RType RTyCon RTyVar       -- ^ "Resolved" version
-type RRep        = RTypeRep RTyCon RTyVar
-type BSort       = BRType    ()
-type RSort       = RRType    ()
-type BPVar       = PVar      BSort
-type RPVar       = PVar      RSort
-type RReft       = UReft     F.Reft
-type BareType    = BRType    RReft
-type SpecType    = RRType    RReft
-type SpecRep     = RRep      RReft
-type SpecProp    = RRProp    RReft
-type RRProp r    = Ref       RSort (RRType r)
-type BRProp r    = Ref       BSort (BRType r)
-type SpecRTVar   = RTVar     RTyVar RSort
-
-
-
-type LocBareType = F.Located BareType
-type LocSpecType = F.Located SpecType
-
-type SpecRTEnv   = RTEnv RTyVar SpecType
-type BareRTEnv   = RTEnv Symbol BareType
-type BareRTAlias = RTAlias Symbol BareType
-type SpecRTAlias = RTAlias RTyVar SpecType
-
-
-class SubsTy tv ty a where
-  subt :: (tv, ty) -> a -> a
-
-class (Eq c) => TyConable c where
-  isFun    :: c -> Bool
-  isList   :: c -> Bool
-  isTuple  :: c -> Bool
-  ppTycon  :: c -> Doc
-  isClass  :: c -> Bool
-  isEmbeddedDict :: c -> Bool
-  isEqual  :: c -> Bool
-  isOrdCls  :: c -> Bool
-  isEqCls   :: c -> Bool
-
-  isNumCls  :: c -> Bool
-  isFracCls :: c -> Bool
-
-  isClass   = const False
-  isEmbeddedDict c = isNumCls c || isEqual c || isOrdCls c || isEqCls c
-  isOrdCls  = const False
-  isEqCls   = const False
-  isEqual   = const False
-  isNumCls  = const False
-  isFracCls = const False
-
-
--- Should just make this a @Pretty@ instance but its too damn tedious
--- to figure out all the constraints.
-
-type OkRT c tv r = ( TyConable c
-                   , F.PPrint tv, F.PPrint c, F.PPrint r
-                   , Reftable r, Reftable (RTProp c tv ()), Reftable (RTProp c tv r)
-                   , Eq c, Eq tv
-                   , Hashable tv
-                   )
-
--------------------------------------------------------------------------------
--- | TyConable Instances -------------------------------------------------------
--------------------------------------------------------------------------------
-
-instance TyConable RTyCon where
-  isFun      = isArrowTyCon . rtc_tc
-  isList     = (listTyCon ==) . rtc_tc
-  isTuple    = Ghc.isTupleTyCon   . rtc_tc
-  isClass    = isClass . rtc_tc -- isClassRTyCon
-  isEqual    = isEqual . rtc_tc
-  ppTycon    = F.toFix
-
-  isNumCls c  = maybe False (isClassOrSubClass isNumericClass)
-                (tyConClass_maybe $ rtc_tc c)
-  isFracCls c = maybe False (isClassOrSubClass isFractionalClass)
-                (tyConClass_maybe $ rtc_tc c)
-  isOrdCls  c = maybe False isOrdClass (tyConClass_maybe $ rtc_tc c)
-  isEqCls   c = isEqCls (rtc_tc c)
-
-
-instance TyConable TyCon where
-  isFun      = isArrowTyCon
-  isList     = (listTyCon ==)
-  isTuple    = Ghc.isTupleTyCon
-  isClass c  = isClassTyCon c   || isEqual c -- c == eqPrimTyCon
-  isEqual c  = c == eqPrimTyCon || c == eqReprPrimTyCon
-  ppTycon    = text . showPpr
-
-  isNumCls c  = maybe False (isClassOrSubClass isNumericClass)
-                (tyConClass_maybe c)
-  isFracCls c = maybe False (isClassOrSubClass isFractionalClass)
-                (tyConClass_maybe c)
-  isOrdCls c  = maybe False isOrdClass
-                (tyConClass_maybe c)
-  isEqCls  c  = isPrelEqTyCon c
-
-isClassOrSubClass :: (Class -> Bool) -> Class -> Bool
-isClassOrSubClass p cls
-  = p cls || any (isClassOrSubClass p . fst)
-                 (mapMaybe getClassPredTys_maybe (classSCTheta cls))
-
--- MOVE TO TYPES
-instance TyConable Symbol where
-  isFun   s = F.funConName == s
-  isList  s = F.listConName == s
-  isTuple s = F.tupConName == s
-  ppTycon   = text . F.symbolString
-
-instance TyConable F.LocSymbol where
-  isFun   = isFun   . F.val
-  isList  = isList  . F.val
-  isTuple = isTuple . F.val
-  ppTycon = ppTycon . F.val
-
-instance TyConable BTyCon where
-  isFun   = isFun . btc_tc
-  isList  = isList . btc_tc
-  isTuple = isTuple . btc_tc
-  isClass = isClassBTyCon
-  ppTycon = ppTycon . btc_tc
-
-
-instance Eq RTyCon where
-  x == y = rtc_tc x == rtc_tc y
-
-instance Eq BTyCon where
-  x == y = btc_tc x == btc_tc y
-
-instance Ord BTyCon where
-  compare x y = compare (btc_tc x) (btc_tc y)
-
-instance F.Fixpoint RTyCon where
-  toFix (RTyCon c _ _) = text $ showPpr c
-
-instance F.Fixpoint BTyCon where
-  toFix = text . F.symbolString . F.val . btc_tc
-
-instance F.Fixpoint Cinfo where
-  toFix = text . showPpr . ci_loc
-
-instance Show Cinfo where
-  show = show . F.toFix
-
-instance F.PPrint RTyCon where
-  pprintTidy k c
-    | ppDebug ppEnv = F.pprintTidy k tc  <-> angleBrackets (F.pprintTidy k pvs)
-    | otherwise     = text . showPpr . rtc_tc $ c
-    where
-      tc            = F.symbol (rtc_tc c)
-      pvs           = rtc_pvars c
-
-instance F.PPrint BTyCon where
-  pprintTidy _ = text . F.symbolString . F.val . btc_tc
-
-instance F.PPrint v => F.PPrint (RTVar v s) where
-  pprintTidy k (RTVar x _) = F.pprintTidy k x
-
-instance Show RTyCon where
-  show = F.showpp
-
-instance Show BTyCon where
-  show = F.showpp
-
-instance F.Loc BTyCon where
-  srcSpan = F.srcSpan . btc_tc
 
 --------------------------------------------------------------------------------
 -- | Refined Instances ---------------------------------------------------------
@@ -1172,140 +434,6 @@ instance Show (Axiom Var Type CoreExpr) where
                                          "\nRHS      :" ++ showPpr rhs
 
 --------------------------------------------------------------------------------
--- | Data type refinements
---------------------------------------------------------------------------------
-data DataDecl   = DataDecl
-  { tycName   :: DataName              -- ^ Type  Constructor Name
-  , tycTyVars :: [Symbol]              -- ^ Tyvar Parameters
-  , tycPVars  :: [PVar BSort]          -- ^ PVar  Parameters
-  , tycDCons  :: Maybe [DataCtor]      -- ^ Data Constructors (Nothing is reserved for non-GADT style empty data declarations)
-  , tycSrcPos :: !F.SourcePos          -- ^ Source Position
-  , tycSFun   :: Maybe SizeFun         -- ^ Default termination measure
-  , tycPropTy :: Maybe BareType        -- ^ Type of Ind-Prop
-  , tycKind   :: !DataDeclKind         -- ^ User-defined or Auto-lifted
-  } deriving (Data, Typeable, Generic)
-    deriving Hashable via Generically DataDecl
-
--- | The name of the `TyCon` corresponding to a `DataDecl`
-data DataName
-  = DnName !F.LocSymbol                -- ^ for 'isVanillyAlgTyCon' we can directly use the `TyCon` name
-  | DnCon  !F.LocSymbol                -- ^ for 'FamInst' TyCon we save some `DataCon` name
-  deriving (Eq, Ord, Data, Typeable, Generic)
-
--- | Data Constructor
-data DataCtor = DataCtor
-  { dcName   :: F.LocSymbol            -- ^ DataCon name
-  , dcTyVars :: [F.Symbol]             -- ^ Type parameters
-  , dcTheta  :: [BareType]             -- ^ The GHC ThetaType corresponding to DataCon.dataConSig
-  , dcFields :: [(Symbol, BareType)]   -- ^ field-name and field-Type pairs
-  , dcResult :: Maybe BareType         -- ^ Possible output (if in GADT form)
-  } deriving (Data, Typeable, Generic, Eq)
-    deriving Hashable via Generically DataCtor
-
--- | Termination expressions
-data SizeFun
-  = IdSizeFun              -- ^ \x -> F.EVar x
-  | SymSizeFun F.LocSymbol -- ^ \x -> f x
-  deriving (Data, Typeable, Generic, Eq)
-  deriving Hashable via Generically SizeFun
-
--- | What kind of `DataDecl` is it?
-data DataDeclKind
-  = DataUser           -- ^ User defined data-definitions         (should have refined fields)
-  | DataReflected      -- ^ Automatically lifted data-definitions (do not have refined fields)
-  deriving (Eq, Data, Typeable, Generic, Show)
-  deriving Hashable via Generically DataDeclKind
-
-instance Show SizeFun where
-  show IdSizeFun      = "IdSizeFun"
-  show (SymSizeFun x) = "SymSizeFun " ++ show (F.val x)
-
-szFun :: SizeFun -> Symbol -> Expr
-szFun IdSizeFun      = F.EVar
-szFun (SymSizeFun f) = \x -> F.mkEApp (F.symbol <$> f) [F.EVar x]
-
-data HasDataDecl
-  = NoDecl  (Maybe SizeFun)
-  | HasDecl
-  deriving (Show)
-
-instance F.PPrint HasDataDecl where
-  pprintTidy _ HasDecl    = text "HasDecl"
-  pprintTidy k (NoDecl z) = text "NoDecl" <+> parens (F.pprintTidy k z)
-
-hasDecl :: DataDecl -> HasDataDecl
-hasDecl d
-  | null (tycDCons d)
-  = NoDecl (tycSFun d)
-  -- // | Just s <- tycSFun d, null (tycDCons d)
-  -- // = NoDecl (Just s)
-  | otherwise
-  = HasDecl
-
-instance Hashable DataName where
-  hashWithSalt i = hashWithSalt i . F.symbol
-
-
-instance NFData   SizeFun
-instance B.Binary SizeFun
-instance NFData   DataDeclKind
-instance B.Binary DataDeclKind
-instance B.Binary DataName
-instance B.Binary DataCtor
-instance B.Binary DataDecl
-
-instance Eq DataDecl where
-  d1 == d2 = tycName d1 == tycName d2
-
-instance Ord DataDecl where
-  compare d1 d2 = compare (tycName d1) (tycName d2)
-
-instance F.Loc DataCtor where
-  srcSpan = F.srcSpan . dcName
-
-instance F.Loc DataDecl where
-  srcSpan = srcSpanFSrcSpan . sourcePosSrcSpan . tycSrcPos
-
-instance F.Loc DataName where
-  srcSpan (DnName z) = F.srcSpan z
-  srcSpan (DnCon  z) = F.srcSpan z
-
-
--- | For debugging.
-instance Show DataDecl where
-  show dd = printf "DataDecl: data = %s, tyvars = %s, sizeFun = %s, kind = %s" -- [at: %s]"
-              (show $ tycName   dd)
-              (show $ tycTyVars dd)
-              (show $ tycSFun   dd)
-              (show $ tycKind   dd)
-
-
-instance Show DataName where
-  show (DnName n) =               show (F.val n)
-  show (DnCon  c) = "datacon:" ++ show (F.val c)
-
-instance F.PPrint SizeFun where
-  pprintTidy _ IdSizeFun      = "[id]"
-  pprintTidy _ (SymSizeFun x) = brackets (F.pprint (F.val x))
-
-instance F.Symbolic DataName where
-  symbol = F.val . dataNameSymbol
-
-instance F.Symbolic DataDecl where
-  symbol = F.symbol . tycName
-
-instance F.PPrint DataName where
-  pprintTidy k (DnName n) = F.pprintTidy k (F.val n)
-  pprintTidy k (DnCon  n) = F.pprintTidy k (F.val n)
-
-  -- symbol (DnName z) = F.suffixSymbol "DnName" (F.val z)
-  -- symbol (DnCon  z) = F.suffixSymbol "DnCon"  (F.val z)
-
-dataNameSymbol :: DataName -> F.LocSymbol
-dataNameSymbol (DnName z) = z
-dataNameSymbol (DnCon  z) = z
-
---------------------------------------------------------------------------------
 -- | Refinement Type Aliases
 --------------------------------------------------------------------------------
 data RTAlias x a = RTA
@@ -1330,6 +458,9 @@ lmapEAlias (LMap v ys e) = F.atLoc v (RTA (F.val v) [] ys e) -- (F.loc v) (F.loc
 --------------------------------------------------------------------------------
 -- | Constructor and Destructors for RTypes ------------------------------------
 --------------------------------------------------------------------------------
+type RRep        = RTypeRep RTyCon RTyVar
+type SpecRep     = RRep      RReft
+
 data RTypeRep c tv r = RTypeRep
   { ty_vars   :: [(RTVar tv (RType c tv ()), r)]
   , ty_preds  :: [PVar (RType c tv ())]
@@ -1754,7 +885,12 @@ foldReft' logicBind bsc g f
               (\_ γ -> γ)
               F.emptySEnv
 
+tyVarIsVal :: RTVar tv s -> Bool
+tyVarIsVal = rtvinfoIsVal . ty_var_info
 
+rtvinfoIsVal :: RTVInfo s -> Bool
+rtvinfoIsVal RTVNoInfo{} = False
+rtvinfoIsVal RTVInfo{..} = rtv_is_val
 
 -- efoldReft :: Reftable r =>(p -> [RType c tv r] -> [(Symbol, a)])-> (RType c tv r -> a)-> (SEnv a -> Maybe (RType c tv r) -> r -> c1 -> c1)-> SEnv a-> c1-> RType c tv r-> c1
 efoldReft :: (Reftable r, TyConable c)
@@ -1916,19 +1052,6 @@ mapRBase _ t                  = t
 -- | F.PPrint -----------------------------------------------------------------
 -----------------------------------------------------------------------------
 
-instance F.PPrint (PVar a) where
-  pprintTidy _ = pprPvar
-
-pprPvar :: PVar a -> Doc
-pprPvar (PV s _ _ xts) = F.pprint s <+> hsep (F.pprint <$> dargs xts)
-  where
-    dargs              = map thd3 . takeWhile (\(_, x, y) -> F.EVar x /= y)
-
-
-instance F.PPrint Predicate where
-  pprintTidy _ (Pr [])  = text "True"
-  pprintTidy k (Pr pvs) = hsep $ punctuate (text "&") (F.pprintTidy k <$> pvs)
-
 
 -- | The type used during constraint generation, used
 --   also to define contexts for errors, hence in this
@@ -2003,7 +1126,7 @@ emptyDiagnostics :: Diagnostics
 emptyDiagnostics = Diagnostics mempty mempty
 
 noErrors :: Diagnostics -> Bool
-noErrors = L.null . dErrors
+noErrors = null . dErrors
 
 allWarnings :: Diagnostics -> [Warning]
 allWarnings = dWarnings
@@ -2046,6 +1169,13 @@ instance NFData Cinfo
 
 instance F.PPrint Cinfo where
   pprintTidy k = F.pprintTidy k . ci_loc
+
+instance F.Fixpoint Cinfo where
+  toFix = text . showPpr . ci_loc
+
+instance Show Cinfo where
+  show = show . F.toFix
+
 --------------------------------------------------------------------------------
 -- | Module Names --------------------------------------------------------------
 --------------------------------------------------------------------------------
@@ -2504,31 +1634,3 @@ instance Eq ctor => Semigroup (MSpec ty ctor) where
 instance Eq ctor => Monoid (MSpec ty ctor) where
   mempty = MSpec M.empty M.empty M.empty []
   mappend = (<>)
-
-
-
---------------------------------------------------------------------------------
--- Nasty PP stuff
---------------------------------------------------------------------------------
-
-instance F.PPrint BTyVar where
-  pprintTidy _ (BTV α) = text (F.symbolString α)
-
-instance F.PPrint RTyVar where
-  pprintTidy k (RTV α)
-   | ppTyVar ppEnv  = F.pprintTidy k (F.symbol α) -- shows full tyvar
-   | otherwise      = ppr_tyvar_short α           -- drops the unique-suffix
-   where
-     ppr_tyvar_short :: TyVar -> Doc
-     ppr_tyvar_short = text . showPpr
-
-instance (F.PPrint r, Reftable r, F.PPrint t, F.PPrint (RType c tv r)) => F.PPrint (Ref t (RType c tv r)) where
-  pprintTidy k (RProp ss s) = ppRefArgs k (fst <$> ss) <+> F.pprintTidy k s
-
-ppRefArgs :: F.Tidy -> [Symbol] -> Doc
-ppRefArgs _ [] = empty
-ppRefArgs k ss = text "\\" <-> hsep (ppRefSym k <$> ss ++ [F.vv Nothing]) <+> "->"
-
-ppRefSym :: (Eq a, IsString a, F.PPrint a) => F.Tidy -> a -> Doc
-ppRefSym _ "" = text "_"
-ppRefSym k s  = F.pprintTidy k s
